starting to load league

+++ SpectatorFootball/League/Loaded_League_Structure.cs
@@ -11,7 +11,7 @@ namespace SpectatorFootball.League
     public class Loaded_League_Structure
     {
         public League_State LState { get; set; }
-        public int Current_Year{ get; set; }
+        public long Current_Year{ get; set; }
         public List<Standings_Row> Standings = null;
         public Season season = null;
